Sorry about all the changes. Should've been better about commiting in more chunks. Added: Images for UI( flip, back, audio on/off, check and X in black and white and star outline / filled) Colors in colors.xml I'm using @color/dark_warm and cool, ../light_warm and cool Color pallete can easily be changed by swapping out the hex value in colors.xml Refactored our review activity and layout. see activity_basic_review.xml Removed some of the buttons we had since they kind of just cluttered the overall lok and I figure if the user wants to swap the book / chapter then we should just have that be a different screen. Added some functions to test changing the button image for the tagged star. Word Bank Activity Thought is to use this as a selection screen for what chapters to review and to let the user see all the words in a given chapter. Added function to toggle displaying the list of chapters below the book name. Audio test is on main activity for just a sanity check on how to load audio and play.
DatabaseManager
Using the Room database I have it set up to work with a simple dataset. I'm currently working on finishing up a test csv file with the needed audio files to populate the database with real data. Heads up, changes to the database can cause the app to crash since it generates a file. I found that I could wipe the data from my emulator and that fixed it. Shouldn't be a big deal once we have the full dataset in. Seems like the correct usage is using version numbering and writing a function to convert up to the new version. I'm lazy and just delete the DB.
Some misc xml files to define looks. Things like adding rounded rectangle backgrounds to image buttons and removing the rounded corners to other buttons in the wordbank view.
activity_basic_review.xml shows the color scheme. Let me know what you think.
Also modified the app so it no longer has that bar on top of everything all the time. Goes a long way in making it look like an actual app.
                                               .--.
                                               `.  \
                                                 \  \
                                                  .  \
                                                  :   .
                                                  |    .
                                                  |    :
                                                  |    |
  ..._  ___                                       |    |
 `."".`''''""--..___                              |    |
 ,-\  \             ""-...__         _____________/    |
 / ` " '                    `""""""""                  .
 \                                                      L
 (>                                                      \
/                                                         \
\_    ___..---.                                            L
  `--'         '.                                           \
                 .                                           \_
                _/`.                                           `.._
             .'     -.                                             `.
            /     __.-Y     /''''''-...___,...--------.._            |
           /   _."    |    /                ' .      \   '---..._    |
          /   /      /    /                _,. '    ,/           |   |
          \_,'     _.'   /              /''     _,-'            _|   |
                  '     /               `-----''               /     |
                  `...-'                                       `...-'